# internal/api/openapi.vodoge.yaml
openapi: 3.0.3
info:
  title: VoDoge Main API
  version: 1.0.0-main
  description: |
    VoDoge 主服务控制面 API。
    - 服务器前缀：`/api`
    - 登录方式：`POST /auth/login`
    - 受保护接口统一使用 `Authorization: Bearer <token>`
    - 文档页面：`GET /docs`
servers:
  - url: /api
    description: VoDoge main API prefix
security:
  - BearerAuth: []
...
paths:
...
  /extensions:
    get:
      tags: [extensions]
      summary: 已安装插件
      responses:
        '200':
          description: 插件列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/InstalledPlugin'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /extensions/{id}:
    put:
      tags: [extensions]
      summary: 启用或禁用插件
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePluginRequest'
      responses:
        '200':
          description: 更新后的插件
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/InstalledPlugin'
        '400':
          $ref: '#/components/responses/ErrorMessage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/ErrorMessage'
    delete:
      tags: [extensions]
      summary: 卸载插件
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 已卸载
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/ErrorMessage'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: SessionToken
...
  responses:
    Unauthorized:
      description: 未授权
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    ErrorMessage:
      description: 错误信息
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  schemas:
    Envelope:
      type: object
      description: |
        所有成功响应（2xx）的信封。

        `data` 恒存在（无资源可返回时为 `null`），`meta` 为空时不出现。
        `data` 与 `error` 互斥且必有其一——判别是结构性的，不再依赖
        `status:"ok"` 那种字符串（它曾经出现在 200 响应里表示失败）。

        **不套信封的三处**：SSE 事件帧（领域事件流，不是 HTTP 响应）、
        `GET /ping`（`/api` 之外的存活探针）、websheet 的承载页与代理通道
        （内容由运营商页面决定）。
      required: [data, request_id]
      properties:
        data:
          nullable: true
          description: 资源本身。集合为数组（空集合是 `[]`），纯动作为 `null`。
        meta:
          type: object
          additionalProperties: true
          description: |
            关于**这次操作**的说明，不是资源的一部分。常见键：

            - `message` 给人看的操作结果
            - `warning` / `warning_code` 操作成功但有保留
            - `requires_restart` / `started` / `applied` 生效状态
            - `device_limit` 集合的额度上限
            - `space_delta` eSIM 空间变化
            - `thread_empty` / `deleted` 删除的副作用
            - `channel` USSD 走的通路（vowifi / cs）
            - `range` 查询参数回显
        request_id:
          type: string
          description: 与服务端访问日志对应，排查时用它定位。
    ErrorEnvelope:
      type: object
      description: 所有错误响应（4xx/5xx）的信封。
      required: [error, request_id]
      properties:
        error:
          $ref: '#/components/schemas/ErrorDetail'
        request_id:
          type: string
    ErrorDetail:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: |
            供程序判别的稳定标识。多数端点用按 HTTP 状态推导的通用码
            （`bad_request` / `not_found` / `conflict` / `internal_error` …），
            **这些不值得分支**，用 HTTP 状态即可。需要客户端据以决策的场景有
            专属码：`ESIM_BUSY`、`ESIM_DOWNLOAD_IN_PROGRESS`、
            `e911_*`、`websheet_*`。
        message:
          type: string
          description: 给人读的说明，不要用于程序分支。
        details:
          type: object
          additionalProperties: true
          description: |
            需要客户端据以决策的结构化数据，与给人读的 message 分开。

            - 409 `ESIM_BUSY`：`busy` / `reason` / `retry_after_ms`（另有 `Retry-After` 头）
            - 409 `ESIM_DOWNLOAD_IN_PROGRESS`：`busy` / `task_id`（可直接订阅该任务）
...
    PluginContribution:
      type: object
      required: [id, label, location, entry]
      properties:
        id:
          type: string
        label:
          type: string
        label_zh:
          type: string
        location:
          type: string
          enum: [sidebar, proxy]
        after:
          type: string
        entry:
          type: string
...
    InstalledPlugin:
      type: object
      required: [id, name, version, contributions, enabled, backend_available, backend_running, installed_at, sha256]
      properties:
        id:
          type: string
        name:
          type: string
        version:
          type: string
        description:
          type: string
        author:
          type: string
        homepage:
          type: string
        permissions:
          type: array
          items:
            type: string
        contributions:
          type: array
          items:
            $ref: '#/components/schemas/PluginContribution'
        enabled:
          type: boolean
        backend_available:
          type: boolean
        backend_running:
          type: boolean
        backend_error:
          type: string
        installed_at:
          type: string
          format: date-time
        sha256:
          type: string
...
    UpdatePluginRequest:
      type: object
      required: [enabled]
      properties:
        enabled:
          type: boolean